renamed branch and changed context files directory
