+++ src/routes/PrivateRoutes.js
@@ -1,6 +1,6 @@
 import React, { useContext, useEffect } from 'react';
 import { Route, withRouter } from 'react-router-dom';
-import { AuthContext } from '../contexts/auth/AuthContext';
+import { AuthContext } from '../components/login/context/AuthContext';
 
 const PrivateRoutes = ({ location, history, ...props }) => {
   const { currentUser } = useContext(AuthContext);
