+++ src/routes/RoutesService.js
@@ -1,6 +1,6 @@
 import React, { useContext, useEffect, useMemo } from 'react';
 import { withRouter } from 'react-router-dom';
-import { AuthContext } from '../contexts/auth/AuthContext';
+import { AuthContext } from '../components/login/context/AuthContext';
 
 function RouteService({ location, history }) {
   const { currentUser } = useContext(AuthContext);
